refactor: update accesability for util functions
