+++ perception_eval/test/matching/test_objects_filter.py
@@ -36,7 +36,7 @@ from perception_eval.matching.objects_filter import get_positive_objects
 from perception_eval.object import DynamicObject
 from perception_eval.result import get_object_results
 from perception_eval.result import PerceptionObjectResult
-from perception_eval.util.debug import get_objects_with_difference
+from perception_eval.util import get_objects_with_difference
 
 
 class TestObjectsFilter(unittest.TestCase):
