+++ perception_eval/test/metrics/tracking/test_clear.py
@@ -30,7 +30,7 @@ from perception_eval.metrics.tracking.clear import CLEAR
 from perception_eval.object import DynamicObject
 from perception_eval.result import get_object_results
 from perception_eval.result import PerceptionObjectResult
-from perception_eval.util.debug import get_objects_with_difference
+from perception_eval.util import get_objects_with_difference
 
 
 class AnswerCLEAR:
