+++ perception_eval/test/metrics/tracking/test_tracking_metrics_score.py
@@ -31,7 +31,7 @@ from perception_eval.metrics.tracking.tracking_metrics_score import TrackingMetr
 from perception_eval.object import DynamicObject
 from perception_eval.result import get_object_results
 from perception_eval.result import PerceptionObjectResult
-from perception_eval.util.debug import get_objects_with_difference
+from perception_eval.util import get_objects_with_difference
 
 
 class TestTrackingMetricsScore(unittest.TestCase):
