+++ perception_eval/test/perception_fp_validation_lsim.py
@@ -26,8 +26,8 @@ from perception_eval.config import PerceptionEvaluationConfig
 from perception_eval.manager import PerceptionEvaluationManager
 from perception_eval.result import get_object_status
 from perception_eval.result import PerceptionFrameConfig
-from perception_eval.util.debug import get_objects_with_difference
-from perception_eval.util.logger_config import configure_logger
+from perception_eval.util import configure_logger
+from perception_eval.util import get_objects_with_difference
 
 if TYPE_CHECKING:
     from perception_eval.object import ObjectType
