+++ perception_eval/test/result/test_object_result.py
@@ -23,7 +23,7 @@ from perception_eval.common.evaluation_task import EvaluationTask
 from perception_eval.object import DynamicObject
 from perception_eval.result import get_object_results
 from perception_eval.result import PerceptionObjectResult
-from perception_eval.util.debug import get_objects_with_difference
+from perception_eval.util import get_objects_with_difference
 
 
 class TestObjectResult(unittest.TestCase):
